fix(ci): python-full-coverage — split oversized test-part files under the 1500-line limit
Six first-party test-part files grew past the 1500-line maintainability
guard this PR. Relocate trailing test functions into new sibling part files
(pure relocation; logic and assertions unchanged), pruning now-unused imports
and dead module-level helpers in each:

  coverage_edges part_001 -> part_013   coverage_edges part_002 -> part_014
  coverage_edges part_004 -> part_015   coverage_edges part_008 -> part_016
  runner part_003 -> part_008           error_paths part_013 -> part_018

pytest-split auto-distributes the new files across coverage shards.
